move auto_device to top-level utils
